Ispravljena istorija transfera
