+++ src/components/layout/ClientHeader.module.css
@@ -1,0 +1,102 @@
+.header {
+  background: var(--navy);
+  padding: 0 2rem;
+  height: 60px;
+  display: flex;
+  align-items: center;
+  gap: 2rem;
+  position: sticky;
+  top: 0;
+  z-index: 100;
+}
+
+.headerBrand { display: flex; align-items: center; gap: 10px; }
+.headerIcon {
+  width: 32px; height: 32px;
+  background: var(--blue);
+  border-radius: 8px;
+  display: flex; align-items: center; justify-content: center;
+}
+.headerIcon svg { width: 16px; height: 16px; }
+.headerBrandText { color: #fff; font-weight: 700; font-size: 16px; }
+
+.headerNav { display: flex; gap: 0.25rem; flex: 1; }
+.headerNavBtn {
+  color: rgba(255,255,255,0.6);
+  background: none;
+  border: none;
+  font-family: var(--font);
+  font-size: 13px;
+  font-weight: 500;
+  padding: 6px 12px;
+  border-radius: 6px;
+  cursor: pointer;
+  transition: all var(--transition);
+  display: flex;
+  align-items: center;
+}
+.headerNavBtn:hover { color: #fff; background: rgba(255,255,255,0.08); }
+.headerNavBtnActive { color: #fff !important; background: rgba(255,255,255,0.12) !important; }
+
+/* ── DROPDOWN ── */
+.payDropdownWrap { position: relative; }
+.payDropdownMenu {
+  position: absolute;
+  top: calc(100% + 8px);
+  left: 0;
+  background: var(--surface);
+  border: 1px solid var(--border);
+  border-radius: var(--radius);
+  box-shadow: var(--shadow);
+  min-width: 190px;
+  z-index: 200;
+  overflow: hidden;
+}
+.payDropdownItem {
+  display: block;
+  width: 100%;
+  text-align: left;
+  background: none;
+  border: none;
+  font-family: var(--font);
+  font-size: 13px;
+  font-weight: 500;
+  color: var(--tx-1);
+  padding: 10px 16px;
+  cursor: pointer;
+  transition: background var(--transition);
+}
+.payDropdownItem:hover { background: var(--bg); color: var(--blue); }
+
+.headerRight { display: flex; align-items: center; gap: 1rem; }
+.headerProfile {
+  display: flex; align-items: center; gap: 8px;
+  color: rgba(255,255,255,0.8);
+  background: none; border: none;
+  font-family: var(--font);
+  font-size: 13px; font-weight: 500;
+  cursor: pointer;
+  padding: 4px 8px;
+  border-radius: 6px;
+  transition: background var(--transition);
+}
+.headerProfile:hover { background: rgba(255,255,255,0.08); }
+.headerAvatar {
+  width: 30px; height: 30px;
+  background: var(--blue);
+  border-radius: 50%;
+  display: flex; align-items: center; justify-content: center;
+  font-size: 11px; font-weight: 700; color: #fff;
+  text-transform: uppercase;
+}
+.headerLogout {
+  display: flex; align-items: center; gap: 6px;
+  background: rgba(255,255,255,0.08);
+  border: 1px solid rgba(255,255,255,0.12);
+  color: rgba(255,255,255,0.7);
+  font-family: var(--font);
+  font-size: 12px; font-weight: 500;
+  padding: 6px 12px; border-radius: 6px;
+  cursor: pointer; transition: all var(--transition);
+}
+.headerLogout:hover { background: rgba(255,255,255,0.15); color: #fff; }
