+++ src/pages/client/ClientDashboard.jsx
@@ -1,13 +1,14 @@
-import { useRef, useLayoutEffect, useState, useEffect } from 'react';
-import { useNavigate }    from 'react-router-dom';
-import gsap               from 'gsap';
-import { clientApi }      from '../../api/endpoints/client';
-import { exchangeApi }    from '../../api/endpoints/exchange';
-import { transfersApi }   from '../../api/endpoints/transfers';
-import { useAuthStore }   from '../../store/authStore';
-import { useFetch }       from '../../hooks/useFetch';
-import Spinner            from '../../components/ui/Spinner';
-import styles             from './ClientDashboard.module.css';
+import { useRef, useLayoutEffect, useState } from 'react';
+import { useNavigate }  from 'react-router-dom';
+import gsap             from 'gsap';
+import { clientApi }    from '../../api/endpoints/client';
+import { exchangeApi }  from '../../api/endpoints/exchange';
+import { transfersApi } from '../../api/endpoints/transfers';
+import { useAuthStore } from '../../store/authStore';
+import { useFetch }     from '../../hooks/useFetch';
+import Spinner          from '../../components/ui/Spinner';
+import ClientHeader     from '../../components/layout/ClientHeader';
+import styles           from './ClientDashboard.module.css';
 
 function formatAmount(amount, currency = 'RSD') {
   return new Intl.NumberFormat('sr-RS', { minimumFractionDigits: 2, maximumFractionDigits: 2 }).format(Math.abs(amount)) + ' ' + currency;
@@ -119,32 +120,15 @@ export default function ClientDashboard() {
   const pageRef  = useRef(null);
   const navigate = useNavigate();
   const user     = useAuthStore(s => s.user);
-  const logout   = useAuthStore(s => s.logout);
 
-  const [selectedAccount,   setSelectedAccount]   = useState(0);
-  const [calcAmount,        setCalcAmount]        = useState('');
-  const [calcFrom,          setCalcFrom]          = useState('EUR');
-  const [calcResult,        setCalcResult]        = useState('');
-  const [showProfile,       setShowProfile]       = useState(false);
-  const [showSwitcher,      setShowSwitcher]      = useState(false);
-  const [paymentRecipient,  setPaymentRecipient]  = useState(null);
-  const [showPayment,       setShowPayment]       = useState(false);
-  const [showPaymentsMenu,  setShowPaymentsMenu]  = useState(false);
-  const [showTransfersMenu, setShowTransfersMenu] = useState(false);
-
-  const paymentsMenuRef  = useRef(null);
-  const transfersMenuRef = useRef(null);
-
-  useEffect(() => {
-    function handleClickOutside(e) {
-      if (paymentsMenuRef.current && !paymentsMenuRef.current.contains(e.target))
-        setShowPaymentsMenu(false);
-      if (transfersMenuRef.current && !transfersMenuRef.current.contains(e.target))
-        setShowTransfersMenu(false);
-    }
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => document.removeEventListener('mousedown', handleClickOutside);
-  }, []);
+  const [selectedAccount,  setSelectedAccount]  = useState(0);
+  const [calcAmount,       setCalcAmount]       = useState('');
+  const [calcFrom,         setCalcFrom]         = useState('EUR');
+  const [calcResult,       setCalcResult]       = useState('');
+  const [showProfile,      setShowProfile]      = useState(false);
+  const [showSwitcher,     setShowSwitcher]     = useState(false);
+  const [paymentRecipient, setPaymentRecipient] = useState(null);
+  const [showPayment,      setShowPayment]      = useState(false);
 
   const clientId = useAuthStore(s => s.user?.client_id ?? s.user?.id);
 
@@ -152,10 +136,9 @@ export default function ClientDashboard() {
     () => clientApi.getAccounts(clientId),
     [clientId]
   );
-  const accounts = Array.isArray(accountsData) ? accountsData : accountsData?.data ?? [];
+  const accounts    = Array.isArray(accountsData) ? accountsData : accountsData?.data ?? [];
   const activeAccount = accounts[selectedAccount];
 
-  // Poslednji transferi — koristi transfersApi koji potvrđeno radi
   const { data: txData, loading: loadingTx } = useFetch(
     () => transfersApi.getHistory(clientId, { page: 1, page_size: 5 }),
     [clientId]
@@ -181,108 +164,9 @@ export default function ClientDashboard() {
     setCalcResult(`${calcAmount} ${calcFrom} = ${(parseFloat(calcAmount) * rate.sell_rate).toFixed(2)} RSD`);
   }
 
-  function handleLogout() { logout(); navigate('/login'); }
-
-  const transfersSubItems = [
-    { label: 'Novi transfer',      path: '/transfers/new' },
-    { label: 'Istorija transfera', path: '/transfers/history' },
-  ];
-
-  const paymentsSubItems = [
-    { label: 'Novo plaćanje',     path: '/client/payments/new' },
-    { label: 'Prenos',            path: '/transfers/new' },
-    { label: 'Primaoci plaćanja', path: '/client/recipients' },
-    { label: 'Pregled plaćanja',  path: '/client/payments' },
-  ];
-
   return (
     <div ref={pageRef} className={styles.page}>
-      <header className={styles.header}>
-        <button className={styles.headerBrand} onClick={() => navigate('/dashboard')} style={{ border: 'none', background: 'transparent', cursor: 'pointer', display: 'flex', alignItems: 'center', gap: '10px' }}>
-          <div className={styles.headerIcon}>
-            <svg viewBox="0 0 24 24" fill="none" stroke="white" strokeWidth="2">
-              <path d="M3 9l9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z"/>
-              <polyline points="9 22 9 12 15 12 15 22"/>
-            </svg>
-          </div>
-          <span className={styles.headerBrandText}>RAFBank</span>
-        </button>
-
-        <nav className={styles.headerNav}>
-          <button className={styles.headerNavBtn} onClick={() => navigate('/client/accounts')}>Računi</button>
-
-          {/* Transferi dropdown */}
-          <div className={styles.payDropdownWrap} ref={transfersMenuRef}>
-            <button
-              className={`${styles.headerNavBtn} ${showTransfersMenu ? styles.headerNavBtnActive : ''}`}
-              onClick={() => setShowTransfersMenu(prev => !prev)}
-            >
-              Transferi
-              <svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" style={{ marginLeft: 4 }}>
-                <polyline points="6 9 12 15 18 9"/>
-              </svg>
-            </button>
-            {showTransfersMenu && (
-              <div className={styles.payDropdownMenu}>
-                {transfersSubItems.map(item => (
-                  <button
-                    key={item.label}
-                    className={styles.payDropdownItem}
-                    onClick={() => { navigate(item.path); setShowTransfersMenu(false); }}
-                  >
-                    {item.label}
-                  </button>
-                ))}
-              </div>
-            )}
-          </div>
-
-          <button className={styles.headerNavBtn} onClick={() => navigate('/client/exchange')}>Menjačnica</button>
-          <button className={styles.headerNavBtn} onClick={() => navigate('/client/cards')}>Kartice</button>
-          <button className={styles.headerNavBtn} onClick={() => navigate('/client/loans')}>Krediti</button>
-
-          {/* Plaćanja dropdown */}
-          <div className={styles.payDropdownWrap} ref={paymentsMenuRef}>
-            <button
-              className={`${styles.headerNavBtn} ${showPaymentsMenu ? styles.headerNavBtnActive : ''}`}
-              onClick={() => setShowPaymentsMenu(prev => !prev)}
-            >
-              Plaćanja
-              <svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" style={{ marginLeft: 4 }}>
-                <polyline points="6 9 12 15 18 9"/>
-              </svg>
-            </button>
-            {showPaymentsMenu && (
-              <div className={styles.payDropdownMenu}>
-                {paymentsSubItems.map(item => (
-                  <button
-                    key={item.label}
-                    className={styles.payDropdownItem}
-                    onClick={() => { navigate(item.path); setShowPaymentsMenu(false); }}
-                  >
-                    {item.label}
-                  </button>
-                ))}
-              </div>
-            )}
-          </div>
-        </nav>
-
-        <div className={styles.headerRight}>
-          <button className={styles.headerProfile} onClick={() => setShowProfile(true)}>
-            <div className={styles.headerAvatar}>{user?.first_name?.[0]}{user?.last_name?.[0]}</div>
-            <span>{user?.first_name} {user?.last_name}</span>
-          </button>
-          <button className={styles.headerLogout} onClick={handleLogout}>
-            <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-              <path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4"/>
-              <polyline points="16 17 21 12 16 7"/>
-              <line x1="21" y1="12" x2="9" y2="12"/>
-            </svg>
-            Odjavi se
-          </button>
-        </div>
-      </header>
+      <ClientHeader onProfileClick={() => setShowProfile(true)} />
 
       <div className={styles.content}>
         <div className={styles.welcome}>
@@ -328,7 +212,7 @@ export default function ClientDashboard() {
             </div>
             {loadingTx ? <Spinner /> : transactions.length === 0 ? (
               <p style={{ color: 'var(--tx-3)', fontSize: 13, textAlign: 'center', padding: '2rem 0' }}>
-                Nema transfera za ovaj račun.
+                Nema transfera za prikaz.
               </p>
             ) : (
               <table className={styles.txTable}>
