+++ src/pages/client/ClientDashboard.module.css
@@ -5,112 +5,6 @@
   flex-direction: column;
 }
 
-/* ── HEADER ── */
-.header {
-  background: var(--navy);
-  padding: 0 2rem;
-  height: 60px;
-  display: flex;
-  align-items: center;
-  gap: 2rem;
-  position: sticky;
-  top: 0;
-  z-index: 100;
-}
-.headerBrand { display: flex; align-items: center; gap: 10px; }
-.headerIcon {
-  width: 32px; height: 32px;
-  background: var(--blue);
-  border-radius: 8px;
-  display: flex; align-items: center; justify-content: center;
-}
-.headerIcon svg { width: 16px; height: 16px; }
-.headerBrandText { color: #fff; font-weight: 700; font-size: 16px; }
-
-.headerNav { display: flex; gap: 0.25rem; flex: 1; }
-.headerNavBtn {
-  color: rgba(255,255,255,0.6);
-  background: none;
-  border: none;
-  font-family: var(--font);
-  font-size: 13px;
-  font-weight: 500;
-  padding: 6px 12px;
-  border-radius: 6px;
-  cursor: pointer;
-  transition: all var(--transition);
-}
-.headerNavBtn:hover { color: #fff; background: rgba(255,255,255,0.08); }
-.headerNavBtnActive { color: #fff !important; background: rgba(255,255,255,0.12) !important; }
-
-/* ── PLAĆANJA DROPDOWN ── */
-.payDropdownWrap {
-  position: relative;
-}
-.payDropdownMenu {
-  position: absolute;
-  top: calc(100% + 8px);
-  left: 0;
-  background: var(--surface);
-  border: 1px solid var(--border);
-  border-radius: var(--radius);
-  box-shadow: var(--shadow);
-  min-width: 190px;
-  z-index: 200;
-  overflow: hidden;
-}
-.payDropdownItem {
-  display: block;
-  width: 100%;
-  text-align: left;
-  background: none;
-  border: none;
-  font-family: var(--font);
-  font-size: 13px;
-  font-weight: 500;
-  color: var(--tx-1);
-  padding: 10px 16px;
-  cursor: pointer;
-  transition: background var(--transition);
-}
-.payDropdownItem:hover {
-  background: var(--bg);
-  color: var(--blue);
-}
-
-.headerRight { display: flex; align-items: center; gap: 1rem; }
-.headerProfile {
-  display: flex; align-items: center; gap: 8px;
-  color: rgba(255,255,255,0.8);
-  background: none; border: none;
-  font-family: var(--font);
-  font-size: 13px; font-weight: 500;
-  cursor: pointer;
-  padding: 4px 8px;
-  border-radius: 6px;
-  transition: background var(--transition);
-}
-.headerProfile:hover { background: rgba(255,255,255,0.08); }
-.headerAvatar {
-  width: 30px; height: 30px;
-  background: var(--blue);
-  border-radius: 50%;
-  display: flex; align-items: center; justify-content: center;
-  font-size: 11px; font-weight: 700; color: #fff;
-  text-transform: uppercase;
-}
-.headerLogout {
-  display: flex; align-items: center; gap: 6px;
-  background: rgba(255,255,255,0.08);
-  border: 1px solid rgba(255,255,255,0.12);
-  color: rgba(255,255,255,0.7);
-  font-family: var(--font);
-  font-size: 12px; font-weight: 500;
-  padding: 6px 12px; border-radius: 6px;
-  cursor: pointer; transition: all var(--transition);
-}
-.headerLogout:hover { background: rgba(255,255,255,0.15); color: #fff; }
-
 /* ── CONTENT ── */
 .content { padding: 2rem; max-width: 1280px; margin: 0 auto; width: 100%; }
 .welcome { margin-bottom: 1.5rem; }
@@ -377,18 +271,8 @@
 .switcherBalance { font-size: 14px; font-weight: 700; color: var(--tx-1); }
 .switcherCheck { color: var(--blue); font-weight: 700; font-size: 16px; }
 
-/* ── COMING SOON ── */
-.comingSoon {
-  padding: 2.5rem 1.5rem;
-  display: flex; flex-direction: column; align-items: center; gap: 0.75rem;
-}
-.comingSoonIcon { font-size: 40px; }
-.comingSoonText { font-size: 16px; font-weight: 600; color: var(--tx-1); }
-.comingSoonSub { font-size: 13px; color: var(--tx-3); margin-bottom: 0.5rem; }
-
 /* ── RESPONSIVE ── */
 @media (max-width: 900px) {
   .grid { grid-template-columns: 1fr; }
-  .headerNav { display: none; }
   .content { padding: 1rem; }
 }
